cloud-run/vector-index/requirements.txt
functions-framework
pillow
sentence_transformers
pinecone-client
algoliasearch